+++ packages/cli/src/config/profileRuntimeApplication.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy CLI boundary retained while larger decomposition continues. */
-
 import { DebugLogger } from '@vybestack/llxprt-code-core';
 import type { Profile } from '@vybestack/llxprt-code-settings';
 import { applyProfileSnapshot } from '@vybestack/llxprt-code-providers/runtime/profileSnapshot.js';
