+++ packages/cli/src/ui/commands/memoryCommand.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy UI boundary retained while larger decomposition continues. */
-
 import {
   getErrorMessage,
   getGlobalCoreMemoryFilePath,
