+++ packages/cli/src/ui/commands/schema/index.ts
@@ -4,8 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy UI boundary retained while larger decomposition continues. */
-
 import type {
   CommandArgumentSchema,
   CompletionResult,
@@ -17,6 +15,12 @@ import type {
 import type { CommandContext } from '../types.js';
 import { filterCompletions } from '../../utils/fuzzyFilter.js';
 import { DebugLogger } from '@vybestack/llxprt-code-core';
+import {
+  normalizeCompletionContext,
+  buildArgumentTokenInfo,
+  resolveActiveStep,
+} from './schemaHelpers.js';
+import type { CompletionInput } from './schemaHelpers.js';
 
 const logger = new DebugLogger('llxprt:ui:schema');
 
@@ -26,234 +30,7 @@ interface FlattenedPath {
   readonly description?: string;
 }
 
-interface CompletionInputObject {
-  args?: string;
-  completedArgs?: readonly string[];
-  partialArg?: string;
-  commandPathLength?: number;
-}
-
-export type CompletionInput = string | CompletionInputObject;
-
-interface NormalizedInput {
-  commandPathLength: number;
-  completedArgs: string[];
-  partialArg: string;
-  hasTrailingSpace: boolean;
-}
-
-interface ValueStepContext {
-  kind: 'value';
-  node: ValueArgument;
-  remainingSchema: CommandArgumentSchema;
-  consumedCount: number;
-  consumedLiterals: number;
-}
-
-interface LiteralStepContext {
-  kind: 'literal';
-  nodes: LiteralArgument[];
-  remainingSchema: CommandArgumentSchema;
-  consumedCount: number;
-  consumedLiterals: number;
-}
-
-interface EmptyContext {
-  kind: 'none';
-  consumedCount: number;
-  consumedLiterals: number;
-}
-
-type ActiveContext = ValueStepContext | LiteralStepContext | EmptyContext;
-
-function mergeSchemas(
-  primary: CommandArgumentSchema | undefined,
-  secondary: CommandArgumentSchema,
-): CommandArgumentSchema {
-  if (!primary || primary.length === 0) {
-    return secondary;
-  }
-  if (secondary.length === 0) {
-    return primary;
-  }
-  return [...primary, ...secondary];
-}
-
-function gatherLiteralGroup(schema: CommandArgumentSchema): {
-  literals: LiteralArgument[];
-  nextIndex: number;
-} {
-  const literals: LiteralArgument[] = [];
-  let index = 0;
-  while (index < schema.length && schema[index]?.kind === 'literal') {
-    literals.push(schema[index] as LiteralArgument);
-    index += 1;
-  }
-  return { literals, nextIndex: index };
-}
-
-function normalizeCompletionContext(
-  input: CompletionInput | undefined,
-  tokenInfo: TokenInfo,
-): NormalizedInput {
-  const tokens = [...tokenInfo.tokens];
-  const hasTrailingSpace = tokenInfo.hasTrailingSpace;
-
-  let commandPathLength = tokens.length > 0 ? 1 : 0;
-  let completedArgs: string[] = [];
-  let partialArg = hasTrailingSpace ? '' : tokenInfo.partialToken;
-  let explicitCompleted = false;
-  let explicitPartial = false;
-
-  if (typeof input === 'object') {
-    if (
-      typeof input.commandPathLength === 'number' &&
-      Number.isFinite(input.commandPathLength)
-    ) {
-      commandPathLength = Math.max(0, Math.floor(input.commandPathLength));
-    }
-
-    if (Array.isArray(input.completedArgs)) {
-      completedArgs = [...input.completedArgs];
-      explicitCompleted = true;
-    }
-
-    if (typeof input.partialArg === 'string') {
-      partialArg = input.partialArg;
-      explicitPartial = true;
-    }
-
-    if (typeof input.args === 'string' && !explicitCompleted) {
-      const trimmed = input.args.trim();
-      completedArgs = trimmed ? trimmed.split(/\s+/) : [];
-    }
-  } else if (typeof input === 'string') {
-    const trimmed = input.trim();
-    completedArgs = trimmed ? trimmed.split(/\s+/) : [];
-  }
-
-  const sanitizedPath = Math.max(0, Math.min(commandPathLength, tokens.length));
-  const argsFromTokens = tokens.slice(sanitizedPath);
-
-  if (explicitCompleted === false) {
-    completedArgs = [...argsFromTokens];
-    if (
-      hasTrailingSpace === false &&
-      tokenInfo.partialToken !== '' &&
-      argsFromTokens.length > 0
-    ) {
-      completedArgs = argsFromTokens.slice(0, -1);
-    }
-  }
-
-  if (!explicitPartial) {
-    partialArg = hasTrailingSpace ? '' : tokenInfo.partialToken;
-    if (tokens.length <= sanitizedPath) {
-      partialArg = '';
-    }
-  }
-
-  return {
-    commandPathLength: sanitizedPath,
-    completedArgs,
-    partialArg,
-    hasTrailingSpace,
-  };
-}
-
-function buildArgumentTokenInfo(
-  completedArgs: readonly string[],
-  partialArg: string,
-  hasTrailingSpace: boolean,
-): TokenInfo {
-  const tokens = [...completedArgs];
-  if (!hasTrailingSpace && partialArg) {
-    tokens.push(partialArg);
-  }
-
-  const positionBase = completedArgs.length;
-  const position = Math.max(1, positionBase + (partialArg ? 2 : 1));
-
-  return {
-    tokens,
-    partialToken: hasTrailingSpace ? '' : partialArg,
-    hasTrailingSpace,
-    position,
-  };
-}
-
-function resolveActiveStep(
-  schema: CommandArgumentSchema,
-  completedArgs: readonly string[],
-): ActiveContext {
-  let currentSchema = schema;
-  const remainingArgs = [...completedArgs];
-  let consumedCount = 0;
-  let consumedLiterals = 0;
-
-  while (currentSchema.length > 0) {
-    const firstNode = currentSchema[0];
-
-    if (firstNode.kind === 'literal') {
-      const { literals, nextIndex } = gatherLiteralGroup(currentSchema);
-
-      if (remainingArgs.length === 0) {
-        return {
-          kind: 'literal',
-          nodes: literals,
-          remainingSchema: currentSchema.slice(nextIndex),
-          consumedCount,
-          consumedLiterals,
-        };
-      }
-
-      const candidate = remainingArgs[0];
-      const matched = literals.find((literal) => literal.value === candidate);
-
-      if (matched) {
-        remainingArgs.shift();
-        consumedCount += 1;
-        consumedLiterals += 1;
-        const remainingSchema =
-          matched.stopPropagation === true
-            ? []
-            : currentSchema.slice(nextIndex);
-        currentSchema = mergeSchemas(matched.next, remainingSchema);
-        continue;
-      }
-
-      return {
-        kind: 'literal',
-        nodes: literals,
-        remainingSchema: currentSchema.slice(nextIndex),
-        consumedCount,
-        consumedLiterals,
-      };
-    }
-
-    const valueNode = firstNode;
-
-    if (remainingArgs.length === 0) {
-      return {
-        kind: 'value',
-        node: valueNode,
-        remainingSchema: currentSchema.slice(1),
-        consumedCount,
-        consumedLiterals,
-      };
-    }
-
-    remainingArgs.shift();
-    consumedCount += 1;
-    currentSchema = mergeSchemas(valueNode.next, currentSchema.slice(1));
-  }
-
-  return {
-    kind: 'none',
-    consumedCount,
-    consumedLiterals,
-  };
-}
+export type { CompletionInput } from './schemaHelpers.js';
 
 async function suggestForValue(
   ctx: CommandContext,
@@ -297,6 +74,20 @@ async function suggestForValue(
  * @plan:PLAN-411-DEEPCOMPLETION
  * @requirement:REQ-002
  */
+function flattenValueOptions(
+  nodePath: readonly string[],
+  nextNode: ValueArgument,
+): FlattenedPath[] {
+  if (!nextNode.options) {
+    return [];
+  }
+  return nextNode.options.map((option) => ({
+    path: [...nodePath, option.value],
+    depth: nodePath.length + 1,
+    description: option.description,
+  }));
+}
+
 function flattenSchemaPaths(
   nodes: LiteralArgument[],
   currentPath: readonly string[] = [],
@@ -309,30 +100,30 @@ function flattenSchemaPaths(
     // If this node has a 'next' that contains value arguments with options,
     // we can create deep paths
     if (node.next && node.next.length > 0) {
-      for (const nextNode of node.next) {
-        // eslint-disable-next-line sonarjs/nested-control-flow -- Existing structure is intentionally preserved; refactoring this boundary is outside the lint slice.
-        if (nextNode.kind === 'value' && nextNode.options) {
-          // For each option in the value node, create a flattened path
-          for (const option of nextNode.options) {
-            flattened.push({
-              path: [...nodePath, option.value],
-              depth: nodePath.length + 1,
-              description: option.description,
-            });
-          }
-        } else if (nextNode.kind === 'literal') {
-          // Recursively flatten nested literals
-          const nestedLiterals = node.next.filter(
-            (n): n is LiteralArgument => n.kind === 'literal',
-          );
-          const nestedPaths = flattenSchemaPaths(nestedLiterals, nodePath);
-          flattened.push(...nestedPaths);
-        }
-      }
+      collectNestedPaths(node.next, nodePath, flattened);
     }
   }
 
   return flattened;
+}
+
+function collectNestedPaths(
+  nextNodes: CommandArgumentSchema,
+  nodePath: readonly string[],
+  flattened: FlattenedPath[],
+): void {
+  for (const nextNode of nextNodes) {
+    if (nextNode.kind === 'value' && nextNode.options) {
+      flattened.push(...flattenValueOptions(nodePath, nextNode));
+    } else if (nextNode.kind === 'literal') {
+      // Recursively flatten nested literals
+      const nestedLiterals = nextNodes.filter(
+        (n): n is LiteralArgument => n.kind === 'literal',
+      );
+      const nestedPaths = flattenSchemaPaths(nestedLiterals, nodePath);
+      flattened.push(...nestedPaths);
+    }
+  }
 }
 
 function suggestForLiterals(
@@ -461,48 +252,62 @@ function inferLiteralHint(nodes: LiteralArgument[]): string {
   return descriptions[0] || 'Select option';
 }
 
-export function tokenize(fullLine: string): TokenInfo {
-  const tokens: string[] = [];
-  let current = '';
-  let inQuotes = false;
-  let escapeNext = false;
-  let hasTrailingSpace = false;
+interface TokenizeState {
+  current: string;
+  inQuotes: boolean;
+  escapeNext: boolean;
+  hasTrailingSpace: boolean;
+}
 
-  // eslint-disable-next-line sonarjs/too-many-break-or-continue-in-loop -- Existing structure is intentionally preserved; refactoring this boundary is outside the lint slice.
-  for (let i = 0; i < fullLine.length; i += 1) {
-    const char = fullLine[i];
-
-    if (escapeNext) {
-      current += char;
-      escapeNext = false;
-      continue;
-    }
-
-    if (char === '\\') {
-      escapeNext = true;
-      continue;
-    }
-
-    if (char === '"' || char === "'") {
-      inQuotes = !inQuotes;
-      continue;
-    }
-
-    if (char === ' ' && !inQuotes) {
-      if (current.length > 0) {
-        tokens.push(current);
-        current = '';
-      }
-      hasTrailingSpace = true;
-      continue;
-    }
-
-    current += char;
-    hasTrailingSpace = false;
+function processTokenChar(
+  char: string,
+  state: TokenizeState,
+  tokens: string[],
+): void {
+  if (state.escapeNext) {
+    state.current += char;
+    state.escapeNext = false;
+    return;
   }
 
-  if (current.length > 0) {
-    tokens.push(current);
+  if (char === '\\') {
+    state.escapeNext = true;
+    return;
+  }
+
+  if (char === '"' || char === "'") {
+    state.inQuotes = !state.inQuotes;
+    return;
+  }
+
+  if (char === ' ' && !state.inQuotes) {
+    if (state.current.length > 0) {
+      tokens.push(state.current);
+      state.current = '';
+    }
+    state.hasTrailingSpace = true;
+    return;
+  }
+
+  state.current += char;
+  state.hasTrailingSpace = false;
+}
+
+export function tokenize(fullLine: string): TokenInfo {
+  const tokens: string[] = [];
+  const state: TokenizeState = {
+    current: '',
+    inQuotes: false,
+    escapeNext: false,
+    hasTrailingSpace: false,
+  };
+
+  for (let i = 0; i < fullLine.length; i += 1) {
+    processTokenChar(fullLine[i], state, tokens);
+  }
+
+  if (state.current.length > 0) {
+    tokens.push(state.current);
   }
 
   const firstToken = tokens.length === 0 ? undefined : tokens[0];
@@ -516,7 +321,7 @@ export function tokenize(fullLine: string): TokenInfo {
     const afterPrefix = firstToken.slice(1);
     if (afterPrefix.length === 0) {
       tokens.shift();
-    } else if (tokens.length > 1 || hasTrailingSpace) {
+    } else if (tokens.length > 1 || state.hasTrailingSpace) {
       tokens[0] = afterPrefix;
     }
   }
@@ -524,7 +329,7 @@ export function tokenize(fullLine: string): TokenInfo {
   let partialTokenValue = '';
   // Stryker disable next-line ConditionalExpression
   const lastToken = tokens.length === 0 ? undefined : tokens[tokens.length - 1];
-  if (!hasTrailingSpace) {
+  if (!state.hasTrailingSpace) {
     const candidateLength = lastToken?.length ?? 0;
     // Stryker disable next-line ConditionalExpression, EqualityOperator
     if (candidateLength > 0 && lastToken) {
@@ -535,7 +340,7 @@ export function tokenize(fullLine: string): TokenInfo {
   return {
     tokens,
     partialToken: partialTokenValue,
-    hasTrailingSpace,
+    hasTrailingSpace: state.hasTrailingSpace,
     position: tokens.length,
   };
 }
