+++ packages/cli/src/ui/commands/setCommand.ts
@@ -4,35 +4,20 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/* eslint-disable complexity, sonarjs/cognitive-complexity, eslint-comments/disable-enable-pair -- Phase 5: legacy UI boundary retained while larger decomposition continues. */
-
 import type {
   SlashCommand,
   CommandContext,
   MessageActionReturn,
 } from './types.js';
 import { CommandKind } from './types.js';
-import type {
-  CommandArgumentSchema,
-  LiteralArgument,
-  TokenInfo,
-  ValueArgument,
-} from './schema/types.js';
 import { getRuntimeApi } from '../contexts/RuntimeContext.js';
 import { coreEvents } from '@vybestack/llxprt-code-core';
 import {
   ephemeralSettingHelp,
   parseEphemeralSettingValue,
 } from '@vybestack/llxprt-code-providers/runtime/ephemeralSettings.js';
-import {
-  getDirectSettingSpecs,
-  resolveAlias,
-} from '@vybestack/llxprt-code-settings';
-import {
-  filterStrings,
-  filterCompletions,
-  getFuzzyEnabled,
-} from '../utils/fuzzyFilter.js';
+import { resolveAlias } from '@vybestack/llxprt-code-settings';
+import { buildSetSchema } from './setCommandSchema.js';
 
 // Subcommand for /set unset - removes ephemeral settings or model parameters
 
@@ -42,321 +27,21 @@ import {
  * - /set <ephemeral-key> <value>
  */
 
-/**
- * Schema-based completion for /set command, redesigned to match P09 test expectations.
- *
- * @plan:PLAN-20251013-AUTOCOMPLETE.P10
- * @requirement:REQ-006
- * @pseudocode ArgumentSchema.md lines 111-130
- * - Line 111: literal `unset`
- * - Line 112: literal `modelparam`
- * - Line 113: literal `emojifilter`
- * - Line 114: nested value arg for param name
- * - Line 115: nested value arg for param value
- * - Line 116: hint for emoji mode
- * - Line 117-120: dynamic completers for providers/params
- *
- * Note: The P09 test uses a mixed approach: literals for 'unset', 'modelparam', 'emojifilter'
- * and a single top-level 'value' argument for other settings. This implementation matches that.
- */
+const setSchema = buildSetSchema();
 
-const toTitleCase = (input: string): string =>
-  input
-    .replace(/([a-z])([A-Z])/g, '$1 $2')
-    .replace(/[-_]/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim()
-    .replace(/\b\w/g, (char) => char.toUpperCase());
-
-type SettingCompleter = NonNullable<ValueArgument['completer']>;
-type SettingLiteralSpec = {
-  value: string;
-  hint: string;
-  description?: string;
-  options?: ReadonlyArray<{ value: string; description?: string }>;
-  completer?: SettingCompleter;
-};
-
-// Common model parameters - used for deep path completion flattening in modelparam context
-const commonParamOptions = [
-  { value: 'temperature', description: 'Sampling temperature (0-2)' },
-  { value: 'max_tokens', description: 'Maximum tokens to generate' },
-  { value: 'top_p', description: 'Nucleus sampling probability' },
-  { value: 'top_k', description: 'Top-k sampling' },
-  { value: 'frequency_penalty', description: 'Frequency penalty (-2 to 2)' },
-  { value: 'presence_penalty', description: 'Presence penalty (-2 to 2)' },
-];
-
-const buildDirectSettingSpecs = (): SettingLiteralSpec[] => {
-  const registrySpecs = getDirectSettingSpecs();
-  return registrySpecs.map(
-    (spec: {
-      value: string;
-      hint: string;
-      description?: string;
-      options?: ReadonlyArray<{ value: string; description?: string }>;
-    }) => ({
-      value: spec.value,
-      hint: spec.hint,
-      description: spec.description,
-      options: spec.options,
-    }),
-  );
-};
-
-const directSettingSpecs: SettingLiteralSpec[] = buildDirectSettingSpecs();
-const directSettingSpecByValue = new Map(
-  directSettingSpecs.map((spec) => [spec.value, spec]),
-);
-
-const createSettingLiteral = (spec: SettingLiteralSpec): LiteralArgument => ({
-  kind: 'literal' as const,
-  value: spec.value,
-  description: spec.description ?? `${toTitleCase(spec.value)} option`,
-  stopPropagation: true,
-  next: [
-    {
-      kind: 'value' as const,
-      name: `${spec.value}-value`,
-      description: spec.description ?? `${toTitleCase(spec.value)} value`,
-      hint: spec.hint,
-      options: spec.options,
-      completer: spec.completer,
-    },
-  ],
-});
-
-const directSettingLiterals = directSettingSpecs.map(createSettingLiteral);
-const directSettingKeys = new Set(directSettingSpecs.map((spec) => spec.value));
-
-const modelParamSchemaNext: CommandArgumentSchema = [
-  {
-    kind: 'value',
-    name: 'param-name',
-    description: 'parameter name',
-    hint: 'parameter name',
-    options: commonParamOptions,
-    completer: async (ctx, partial) => {
-      const enableFuzzy = getFuzzyEnabled(ctx);
-      const modelParams = getRuntimeApi().getActiveModelParams();
-      const paramNames = Object.keys(modelParams);
-      if (paramNames.length > 0) {
-        const matches = filterStrings(paramNames, partial, { enableFuzzy });
-        if (matches.length > 0) {
-          return matches.map((name) => ({
-            value: name,
-            description: `Parameter: ${name}`,
-          }));
-        }
-      }
-
-      return filterCompletions(commonParamOptions, partial, {
-        enableFuzzy,
-      });
-    },
-    next: [
-      {
-        kind: 'value',
-        name: 'param-value',
-        description: 'model parameter value',
-        hint: 'value to set for the parameter (number, string, boolean, or JSON)',
-      },
-    ],
-  },
-];
-
-// Stryker disable StringLiteral -- literal descriptions are static UX copy verified via interaction tests.
-const setSchema: CommandArgumentSchema = [
-  {
-    kind: 'literal',
-    value: 'unset',
-    description: 'Unset option',
-    stopPropagation: true,
-    next: [
-      {
-        kind: 'value',
-        name: 'key',
-        description: 'setting key to remove',
-        completer: async (ctx, partial) => {
-          const ephemeralSettings = getRuntimeApi().getEphemeralSettings();
-          const ephemeralKeys = Object.keys(ephemeralSettings).filter(
-            (key) => ephemeralSettings[key] !== undefined,
-          );
-
-          const specialKeys = [
-            'modelparam',
-            'custom-headers',
-            ...Array.from(directSettingKeys),
-          ];
-          const allKeys = Array.from(
-            new Set([...ephemeralKeys, ...specialKeys]),
-          );
-
-          const enableFuzzy = getFuzzyEnabled(ctx);
-          const filtered = filterStrings(allKeys, partial, { enableFuzzy });
-          return filtered.map((key) => ({
-            value: key,
-            description: `setting: ${key}`,
-          }));
-        },
-        next: [
-          {
-            kind: 'value',
-            name: 'subkey',
-            description: 'nested key for specific settings',
-            hint: async (_ctx, tokens: TokenInfo) => {
-              const key = tokens.tokens[1];
-              if (key === 'modelparam') {
-                return 'model parameter name (e.g., temperature, max_tokens)';
-              }
-              if (key === 'custom-headers') {
-                return 'header name (e.g., Authorization)';
-              }
-              return 'subkey (optional)';
-            },
-            completer: async (ctx, partial, tokens) => {
-              const key = tokens.tokens[1];
-              const enableFuzzy = getFuzzyEnabled(ctx);
-
-              if (key === 'modelparam') {
-                const params = getRuntimeApi().getActiveModelParams();
-                const paramNames = Object.keys(params);
-                const filtered = filterStrings(paramNames, partial, {
-                  enableFuzzy,
-                });
-                return filtered.map((name) => ({
-                  value: name,
-                  description: `Parameter: ${name}`,
-                }));
-              }
-
-              if (key === 'custom-headers') {
-                const headers = getRuntimeApi().getEphemeralSettings()[
-                  'custom-headers'
-                ] as Record<string, string> | undefined;
-                if (headers) {
-                  const headerNames = Object.keys(headers);
-                  const filtered = filterStrings(headerNames, partial, {
-                    enableFuzzy,
-                  });
-                  return filtered.map((name) => ({
-                    value: name,
-                    description: `header: ${name}`,
-                  }));
-                }
-              }
-
-              return [];
-            },
-          },
-        ],
-      },
-    ],
-  },
-  {
-    kind: 'literal',
-    value: 'modelparam',
-    description: 'Model parameter option',
-    stopPropagation: true,
-    next: modelParamSchemaNext,
-  },
-  {
-    kind: 'literal',
-    value: 'emojifilter',
-    description: 'Emoji filter option',
-    stopPropagation: true,
-    next: [
-      {
-        kind: 'value',
-        name: 'mode',
-        description: 'filter mode',
-        hint: 'filter mode',
-        options: directSettingSpecByValue.get('emojifilter')?.options ?? [],
-      },
-    ],
-  },
-  ...directSettingLiterals.filter((literal) => literal.value !== 'emojifilter'),
-  // Stryker disable all -- fallback value handler is exercised through runtime flows and
-  // would require extensive integration scaffolding beyond the schema migration scope.
-  {
-    kind: 'value',
-    name: 'setting',
-    description: 'any ephemeral setting key',
-    options: Object.entries(ephemeralSettingHelp)
-      .filter(([key]) => !directSettingKeys.has(key))
-      .map(([value, description]) => ({
-        value,
-        description,
-      })),
-    next: [
-      {
-        kind: 'value',
-        name: 'setting-value',
-        description: 'setting value',
-        hint: async (_ctx, tokens: TokenInfo) => {
-          const setting = tokens.partialToken || tokens.tokens[0];
-          switch (setting) {
-            case 'context-limit':
-              return 'positive integer (e.g., 100000)';
-            case 'compression-threshold':
-              return 'decimal between 0 and 1 (e.g., 0.7)';
-            case 'emojifilter':
-              return 'allowed, auto, warn, or error';
-            case 'streaming':
-              return 'enabled or disabled';
-            case 'socket-timeout':
-              return 'positive integer in milliseconds (e.g., 60000)';
-            case 'socket-keepalive':
-            case 'socket-nodelay':
-              return 'true or false';
-            case 'shell-replacement':
-              return 'allowlist, all, or none';
-            case 'tool-output-truncate-mode':
-              return 'warn, truncate, or sample';
-            case 'maxTurnsPerPrompt':
-              return 'positive integer or -1 (unlimited)';
-            default:
-              return 'value to set';
-          }
-        },
-        completer: async (ctx, partial, tokens) => {
-          const setting = tokens.tokens[0] || tokens.partialToken;
-          const enableFuzzy = getFuzzyEnabled(ctx);
-
-          // Try to get autocomplete suggestions from the registry first
-          const resolvedSetting = resolveAlias(setting);
-          const registryOptions =
-            directSettingSpecByValue.get(resolvedSetting)?.options;
-
-          if (registryOptions && registryOptions.length > 0) {
-            return filterCompletions(registryOptions, partial, { enableFuzzy });
-          }
-
-          // Special case: custom-headers needs runtime lookup
-          if (setting === 'custom-headers') {
-            const headers = getRuntimeApi().getEphemeralSettings()[
-              'custom-headers'
-            ] as Record<string, string> | undefined;
-            if (headers) {
-              const headerNames = Object.keys(headers);
-              const filtered = filterStrings(headerNames, partial, {
-                enableFuzzy,
-              });
-              return filtered.map((name) => ({
-                value: name,
-                description: `header: ${name}`,
-              }));
-            }
-          }
-
-          return [];
-        },
-      },
-    ],
-  },
-  // Stryker restore all
-];
-// Stryker restore StringLiteral
+function formatParsedValue(value: unknown): string {
+  if (typeof value === 'string') {
+    return value;
+  }
+  if (
+    typeof value === 'number' ||
+    typeof value === 'boolean' ||
+    value === null
+  ) {
+    return String(value);
+  }
+  return JSON.stringify(value);
+}
 
 function handleSetModelParam(parts: string[]): MessageActionReturn {
   if (parts.length < 3) {
@@ -383,16 +68,7 @@ function handleSetModelParam(parts: string[]): MessageActionReturn {
     };
   }
 
-  const formattedValue =
-    // eslint-disable-next-line sonarjs/expression-complexity -- Existing structure is intentionally preserved; refactoring this boundary is outside the lint slice.
-    typeof parsedParamValue === 'string'
-      ? parsedParamValue
-      : // eslint-disable-next-line sonarjs/no-nested-conditional -- Existing structure is intentionally preserved; refactoring this boundary is outside the lint slice.
-        typeof parsedParamValue === 'number' ||
-          typeof parsedParamValue === 'boolean' ||
-          parsedParamValue === null
-        ? String(parsedParamValue)
-        : JSON.stringify(parsedParamValue);
+  const formattedValue = formatParsedValue(parsedParamValue);
 
   return {
     type: 'message',
@@ -433,7 +109,7 @@ function handleUnsetModelParam(subKey: string): MessageActionReturn {
 function handleUnsetCustomHeader(
   targetKey: string,
   subKey: string,
-): MessageActionReturn | null {
+): MessageActionReturn {
   const runtime = getRuntimeApi();
   const currentHeaders = runtime.getEphemeralSettings()['custom-headers'] as
     | Record<string, unknown>
@@ -487,13 +163,7 @@ function handleSetUnset(parts: string[]): MessageActionReturn {
   }
 
   if (resolvedTargetKey === 'custom-headers' && subKey) {
-    return (
-      handleUnsetCustomHeader(targetKey, subKey) ?? {
-        type: 'message',
-        messageType: 'info',
-        content: `No custom header named '${subKey}' found`,
-      }
-    );
+    return handleUnsetCustomHeader(targetKey, subKey);
   }
 
   runtime.setEphemeralSetting(targetKey, undefined);
@@ -615,11 +285,9 @@ export const setCommand: SlashCommand = {
  * Parse a string value into the appropriate type.
  * Handles numbers, booleans, and JSON objects/arrays.
  */
-function parseValue(value: string): unknown {
+export function parseValue(value: string): unknown {
   // Try to parse as number
-  // Static regex for numeric validation - no dynamic parts
-  // eslint-disable-next-line sonarjs/regular-expr
-  if (/^-?\d+(\.\d+)?$/.test(value)) {
+  if (looksNumeric(value)) {
     const num = Number(value);
     if (!isNaN(num)) {
       return num;
@@ -642,3 +310,29 @@ function parseValue(value: string): unknown {
     return value;
   }
 }
+
+function looksNumeric(value: string): boolean {
+  let i = 0;
+  if (value[0] === '-') {
+    i = 1;
+  }
+  const digitsStart = i;
+  while (i < value.length && value[i] >= '0' && value[i] <= '9') {
+    i++;
+  }
+  const hasIntegerDigits = i > digitsStart;
+
+  if (i < value.length && value[i] === '.') {
+    i++;
+    const fractionStart = i;
+    while (i < value.length && value[i] >= '0' && value[i] <= '9') {
+      i++;
+    }
+    const hasFractionDigits = i > fractionStart;
+    // Dot must be followed by at least one digit
+    return hasIntegerDigits && hasFractionDigits && i === value.length;
+  }
+
+  return hasIntegerDigits && i === value.length;
+}
+// Stryker restore all
